+++ src/app/clusters/color-control-server/codegen/color-control-server.cpp
